# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Irrigation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /zones/{id}:
    get:
      operationId: getZone
      tags: [zones]
      summary: Get zone details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zone details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
    patch:
      operationId: updateZone
      tags: [zones]
      summary: Update zone configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateZoneBody"
      responses:
        "200":
          description: Updated zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"
  /zones/{id}/moisture-trend:
    get:
      operationId: getZoneMoistureTrend
      tags: [zones]
      summary: Get moisture trend for a zone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: hours
          in: query
          schema:
            type: integer
            default: 24
      responses:
        "200":
          description: Moisture trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoistureTrendPoint"
...
components:
  schemas:
...
    Zone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cropId:
          type: integer
        cropName:
          type: string
        soilTypeId:
          type: integer
        soilTypeName:
          type: string
        currentMoisture:
          type: number
        targetMoistureMin:
          type: number
        targetMoistureMax:
          type: number
        pumpId:
          type: integer
        pumpStatus:
          type: string
          enum: [on, off, override]
        lastIrrigated:
          type: string
          nullable: true
        status:
          type: string
          enum: [optimal, dry, wet, waterlogged, unknown]
      required:
        - id
        - name
        - cropId
        - cropName
        - soilTypeId
        - soilTypeName
        - currentMoisture
        - targetMoistureMin
        - targetMoistureMax
        - pumpId
        - pumpStatus
        - status
    UpdateZoneBody:
      type: object
      properties:
        name:
          type: string
        cropId:
          type: integer
        soilTypeId:
          type: integer
        targetMoistureMin:
          type: number
        targetMoistureMax:
          type: number
    MoistureTrendPoint:
      type: object
      properties:
        timestamp:
          type: string
        moisture:
          type: number
      required: [timestamp, moisture]
...